app/requirements.txt
# Telegram API
telethon

# google drive downloader
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

aiofiles